LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/before-prepare-iframe-parse-error-inline-classic-expected.txt

CONSOLE MESSAGE: SyntaxError: Left side of assignment is not a reference.
 

PASS Sanity check around top-level Window 
PASS Eval: Move parse-error inline classic script to iframe before-prepare 
PASS <script> load: Move parse-error inline classic script to iframe before-prepare 
PASS <script> error: Move parse-error inline classic script to iframe before-prepare 
FAIL window error: Move parse-error inline classic script to iframe before-prepare assert_unreached: Window error event shouldn't fired on destination window Reached unreachable code
